requirements.txt
discord.py==1.7.3
rich==12.6.0